% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{build_covariates_description}
\alias{build_covariates_description}
\title{build_covariates_description}
\usage{
build_covariates_description(
  cov_names,
  cov_labels,
  cov_units,
  cov_types,
  cov_values,
  cov_descriptions
)
}
\arguments{
\item{cov_names}{Character vector of covariate names}

\item{cov_labels}{Character vector of covariate labels}

\item{cov_units}{Character vector of covariate units}

\item{cov_types}{Character vector of covariate types}

\item{cov_values}{List of covariate value ranges}

\item{cov_descriptions}{Character vector of descriptions}
}
\value{
Covariates structure for description section
}
\description{
Build covariates structure for model description section
}
