% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{build_error_structure}
\alias{build_error_structure}
\title{build_error_structure}
\usage{
build_error_structure(
  type = "additive",
  initial_value = 0.1,
  c0 = 0,
  c1 = 0.1,
  c2 = 0,
  c3 = 0
)
}
\arguments{
\item{type}{Error type (additive, proportional, gamma)}

\item{initial_value}{Initial error value}

\item{c0}{Coefficient c0}

\item{c1}{Coefficient c1}

\item{c2}{Coefficient c2}

\item{c3}{Coefficient c3}
}
\value{
Error structure
}
\description{
Build error model structure for model JSON
}
