% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{build_primary_parameters}
\alias{build_primary_parameters}
\title{build_primary_parameters}
\usage{
build_primary_parameters(param_names, param_types, param_mins, param_maxs)
}
\arguments{
\item{param_names}{Character vector of parameter names}

\item{param_types}{Character vector of parameter types (ab, fx, etc.)}

\item{param_mins}{Numeric vector of minimum values}

\item{param_maxs}{Numeric vector of maximum values}
}
\value{
Primary parameters structure
}
\description{
Build primary parameters structure for model JSON
}
