% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{build_route_structure}
\alias{build_route_structure}
\title{build_route_structure}
\usage{
build_route_structure(routes, compartments = NULL)
}
\arguments{
\item{routes}{Character vector of routes}

\item{compartments}{Optional named list mapping routes to compartments}
}
\value{
Route structure for JSON
}
\description{
Build route structure for model JSON
}
