% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{lit_sim}
\alias{lit_sim}
\title{Launch the Literature Simulator app}
\usage{
lit_sim(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Logical, passed to [shiny::runApp()].}

\item{...}{arguments to pass to golem options.}
}
\description{
Launch the Literature Simulator app
}
