% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_compartment_definition}
\alias{parse_compartment_definition}
\title{parse_compartment_definition}
\usage{
parse_compartment_definition(compartment)
}
\arguments{
\item{compartment}{Compartment definition from model JSON}
}
\value{
List with number, central, peripheral fields
}
\description{
Parse compartment definition from model description.
Supports both legacy integer format and new object format:
{ number, central, peripheral }.
}
