% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_covariates_description}
\alias{parse_covariates_description}
\title{parse_covariates_description}
\usage{
parse_covariates_description(covariates)
}
\arguments{
\item{covariates}{Covariates list from model description}
}
\value{
List with structured covariate data
}
\description{
Parse covariates from model description
}
