% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_error_model}
\alias{parse_error_model}
\title{parse_error_model}
\usage{
parse_error_model(error)
}
\arguments{
\item{error}{Error list from model JSON}
}
\value{
List with error model details
}
\description{
Parse error model from model structure
}
