% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_model_covariates}
\alias{parse_model_covariates}
\title{parse_model_covariates}
\usage{
parse_model_covariates(covariates)
}
\arguments{
\item{covariates}{Covariates from model section}
}
\value{
Named list with interpolation methods
}
\description{
Parse covariate interpolation settings from model section
}
