% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_model_equations}
\alias{parse_model_equations}
\title{parse_model_equations}
\usage{
parse_model_equations(equation)
}
\arguments{
\item{equation}{Equation list from model JSON}
}
\value{
Character vector of equations
}
\description{
Parse differential equations from model structure
}
