% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_output_equations}
\alias{parse_output_equations}
\title{parse_output_equations}
\usage{
parse_output_equations(out)
}
\arguments{
\item{out}{Output list from model JSON}
}
\value{
Character vector of output equations
}
\description{
Parse output equations from model structure
}
