% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_primary_parameters}
\alias{parse_primary_parameters}
\title{parse_primary_parameters}
\usage{
parse_primary_parameters(primary)
}
\arguments{
\item{primary}{Primary parameter list from model JSON}
}
\value{
Data frame with parameter info
}
\description{
Parse primary parameters from model structure
}
