% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_route_structure}
\alias{parse_route_structure}
\title{parse_route_structure}
\usage{
parse_route_structure(route)
}
\arguments{
\item{route}{Route data from model JSON}
}
\value{
List with route names and compartment mappings
}
\description{
Parse route from model - can be string or array of objects
}
