% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_bestdose.R
\name{parse_secondary_parameters}
\alias{parse_secondary_parameters}
\title{parse_secondary_parameters}
\usage{
parse_secondary_parameters(secondary)
}
\arguments{
\item{secondary}{Secondary parameter list from model JSON}
}
\value{
Named list of equations
}
\description{
Parse secondary parameter equations from model structure
}
