% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Logical, passed to [shiny::runApp()].}

\item{...}{arguments to pass to golem options.}
}
\description{
Run the Shiny Application
}
